requirements-api.txt
# Production API Requirements (Optimized)
# For Vercel Serverless deployment

# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20

# LangChain & LLM
langchain==0.3.7
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-community==0.3.7

# Vector Store & Embeddings
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.0

# Environment
python-dotenv==1.0.1

# Vercel KV for serverless session storage
vercel-kv==0.3.3
